Make approbation to run against vegaprotocol/specs
- Update pipelineApprobation.groovy to rename variables
- Update pipelineDefaults to use new variables & default to specs

vegaprotocol/specs-internal is being made public, replacing the old
vegaprotocol/specs repo. The latter was just a mirror of the former with
some files filtered out - we no longer feel it is required, so we're
just replacing one with the other. This PR updates the variable names
and defaults related to the repo that approbation runs against.